requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
jinja2==3.1.3
python-multipart==0.0.9
openai==1.75.0
python-dotenv==1.0.1
pandas==2.2.3
tiktoken==0.6.0
langchain>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.0.2
mysql-connector-python>=8.0.0
requests>=2.31.0
sqlalchemy
psycopg2-binary